+++ src/config.rs
@@ -4,6 +4,8 @@ pub struct Config {
     pub sync: bool,
     pub table_size: usize,
     pub bits_per_key: usize,
+    pub compaction_threshold: usize,
+    pub compaction_size_ratio: f64,
 }
 
 impl Default for Config {
@@ -13,6 +15,8 @@ impl Default for Config {
             sync: true,
             table_size: 4096 * 1024,
             bits_per_key: 10,
+            compaction_threshold: 4,
+            compaction_size_ratio: 1.5,
         }
     }
 }